refactor(udp-protocol): move announce types to announce module
